debugging weird jump in trajectory when robot is 0-init

+++ landmark.cpp
@@ -4,14 +4,14 @@
 #include <Eigen/Dense>
 
 Landmark::Landmark() : x(0.0), y(0.0) {}
-Landmark::Landmark(float _x, float _y) : x(_x), y(_y), position({_x, _y}) {}
+Landmark::Landmark(double _x, double _y) : x(_x), y(_y), position({_x, _y}) {}
 Landmark::Landmark(Eigen::Vector2d _position) : position(_position), x(_position.x()), y(_position.y()) {}
 
 // Randomize a pre-initialized Landmark
-void Landmark::randomize(const float &lower_bound, const float &upper_bound) {
+void Landmark::randomize(const double &lower_bound, const double &upper_bound) {
     static std::random_device r;
     static std::mt19937 generate(r());
-    std::uniform_real_distribution<float> P(lower_bound, upper_bound);
+    std::uniform_real_distribution<double> P(lower_bound, upper_bound);
     x = P(generate);
     y = P(generate);
     position = Eigen::Vector2d({x, y});
@@ -39,8 +39,8 @@ bool operator>=(const Landmark& a, const Landmark& b) {
 }
 
 bool operator==(const Landmark& a, const Landmark& b) { 
-    float eps = 1e-4;
-    float dx = std::abs(a.x - b.x);
-    float dy = std::abs(a.y - b.y);
+    double eps = 1e-4;
+    double dx = std::abs(a.x - b.x);
+    double dy = std::abs(a.y - b.y);
     return std::tie(dx, dy) < std::tie(eps, eps);
 }
