+++ utils.h
@@ -3,7 +3,7 @@
 #include <random>
 #include "control.h"
 #include "pose.h"
-double to_radians(float deg);
+double to_radians(double deg);
 double sind(double x);
 double cosd(double x);
 bool isclose(double a, double b, double rtol = 1e-5, double atol = 1e-8);
